changed temperatures to be 2 bytes each, and in units of C
also changed integer types so suspension sensors can maybe be negative now

+++ Inc/sensorFunctions.h
@@ -5,12 +5,14 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
+
 
 
 //function prototypes
-int INFKL800(int adc_value);
-int linearPot750mm12V(int adc_value);
-int RSC28xxx3621x_frontSuspension(int adc_value, int offsetAngle);
+uint16_t INFKL800(int adc_value);
+int8_t linearPot750mm12V(int adc_value);
+int8_t RSC28xxx3621x_frontSuspension(int adc_value, int offsetAngle);
